refactor: extract node shell-state lifecycle into src/core/graph/nodeShell
Move the app-owned node shell-state lifecycle coordination out of the
litegraph folder into an app-layer module, leaving thin call sites in
litegraph. No behaviour change.

- createNodeShellState / setTrackedNodeState / registerNodeState /
  unregisterNodeState / unregisterAllNodeStates now live in
  src/core/graph/nodeShell/nodeShellState.ts
- attachNodeToStores / releaseGraphStores in nodeShellLifecycle.ts are the
  single calls LGraph.add and LGraph.clear make into the app stores
- LGraphNode no longer imports useNodeDataStore
- the litegraph barrel no longer re-exports the registration functions

Amp-Thread-ID: https://ampcode.com/threads/T-01a027af-1c32-7532-84f9-b94aa90620fa
