+++ src/stores/nodeDataStore.test.ts
@@ -3,8 +3,8 @@ import { setActivePinia } from 'pinia'
 import { assert, beforeEach, describe, expect, it } from 'vitest'
 import { computed } from 'vue'
 
+import { transferReplacementOwnership } from '@/core/graph/nodeShell/nodeShellState'
 import { LGraph, LGraphNode } from '@/lib/litegraph/src/litegraph'
-import { transferReplacementOwnership } from '@/lib/litegraph/src/LGraphNode'
 import { toOwningGraphId, toRootGraphId } from '@/types/graphScopeId'
 import type { GraphScope } from '@/types/graphScopeId'
 import type { NodeState } from '@/types/nodeState'
